01_Agent/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangChain和LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.2
langchain-core>=0.1.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# 类型提示
typing-extensions>=4.8.0

# 工具库
httpx>=0.25.0

